Clean up Artisan command to update currencies exchange rates

+++ app/Console/Commands/LoadExchangeRates.php
@@ -23,18 +23,37 @@ class LoadExchangeRates extends Command
      */
     protected $description = 'Load exchange rates for all available currencies';
 
+    const SUCCESS = 0;
+    const ERROR_CODE_MISSING_API_KEY = 1;
+    const ERROR_CODE_API_REQUESTS_FAILED = 1;
+
     /**
      * Execute the console command.
      */
     public function handle()
     {
+        if (! config('services.exchangerate.key')) {
+            $this->error('No Exchangerate-API key available. Rates were not imported');
+            return self::ERROR_CODE_MISSING_API_KEY;
+        }
+
         CurrencyExchangeRate::truncate();
 
         $currencies = Currency::all();
 
+        $errors = 0;
         foreach ($currencies as $fromCurrency) {
-            $exchangeRatesData = Http::get('https://v6.exchangerate-api.com/v6/6680e39093b150914b8c3d21/latest/'.$fromCurrency->code);
-            $convertionRates = $exchangeRatesData->json()['conversion_rates'];
+            try {
+                $exchangeRatesData = Http::get('https://v6.exchangerate-api.com/v6/'.config('services.exchangerate.key').'/latest/'.$fromCurrency->code)
+                    ->throw()
+                    ->json();
+            } catch (\Exception $e) {
+                $this->error('Failed importing exception rates for ' . $fromCurrency->code . ': ' . $e->getMessage());
+                $errors++;
+                continue;
+            }
+
+            $convertionRates = $exchangeRatesData['conversion_rates'];
             foreach ($currencies as $toCurrency) {
                 CurrencyExchangeRate::create([
                     'from_currency_id' => $fromCurrency->id,
@@ -42,8 +61,10 @@ class LoadExchangeRates extends Command
                     'rate' => $convertionRates[$toCurrency->code],
                 ]);
             }
+
+            $this->info('Exchange rates for '.$fromCurrency->code.' loaded successfully');
         }
 
-        // TODO: add some extra messages, then schedule and pass it on to Sentry
+        return $errors == 0 ? self::SUCCESS : self::ERROR_CODE_API_REQUESTS_FAILED;
     }
 }
